refactor(types): reorganize and consolidate schema definitions for clarity
